backend/requirements.txt
django 
djangorestframework 
mysqlclient
djangorestframework-simplejwt
python-dotenv